fix(build): make legacy Setuptools tool compliant with PEP 517 specification
