+++ incipyt/tools/__init__.py
@@ -3,5 +3,5 @@
 from incipyt.tools.base import Tool
 from incipyt.tools.git import Git
 from incipyt.tools.license import License
-from incipyt.tools.setuptools import Setuptools
+from incipyt.tools.pep517 import Setuptools
 from incipyt.tools.venv import Venv
