+++ incipyt/tools/pep517/__init__.py
@@ -1,3 +1,4 @@
 """TO-DO."""
 
 from incipyt.tools.pep517.base import BuildSystem
+from incipyt.tools.pep517.setuptools import Setuptools
